File organisation and implementation of modules for the surveyJS content.
