+++ diagnosis-quiz.html
@@ -15,7 +15,7 @@
     <script type="text/javascript" src="https://unpkg.com/survey-core/survey.core.min.js"></script>
     <script type="text/javascript" src="https://unpkg.com/survey-js-ui/survey-js-ui.min.js"></script>
 
-    <script type="text/javascript" src="javascript/diagnosis.js"></script>
+    <script type="module" src="javascript/diagnosisMain.js"></script>
 
     <title>BeGONEia - Plant Diagnosis</title>
