+++ javascript/modules/surveyTheme.js
@@ -1,0 +1,237 @@
+const coffeeTheme = {
+    "themeName": "coffee",
+    "colorPalette": "light",
+    "isPanelless": false,
+    "backgroundImage": "",
+    "backgroundOpacity": 1,
+    "backgroundImageAttachment": "scroll",
+    "backgroundImageFit": "cover",
+    "cssVariables": {
+        "--sjs-editorpanel-backcolor": "rgba(242, 174, 114, 1)",
+        "--sjs-editorpanel-hovercolor": "rgba(220, 83, 4, 1)",
+        "--sjs-questionpanel-backcolor": "rgba(229, 221, 210, 1)",
+        "--sjs-questionpanel-hovercolor": "rgba(242, 174, 114, 0.5)",
+        "--sjs-questionpanel-cornerRadius": "10px",
+        "--sjs-corner-radius": "10px",
+        "--sjs-base-unit": "8px",
+        "--sjs-font-pagetitle-family": "Brush Script MT, cursive",
+        "--sjs-font-pagetitle-color": "rgba(92, 37, 19, 1)",
+        "--sjs-font-pagedescription-color": "rgba(18, 18, 6, 1)",
+        "--sjs-shadow-small": "0px 0px 0px 3px rgba(92, 37, 19, 1),-5px 5px 0px 5px rgba(191, 73, 4, 0.8)",
+        "--sjs-font-questiontitle-color": "rgba(18, 18, 6, 1)",
+        "--sjs-font-questiondescription-size": "12px",
+        "--sjs-font-questiondescription-color": "rgba(18, 18, 6, 1)",
+        "--sjs-shadow-inner": "0px 0px 0px 2px rgba(92, 37, 19, 1),0px -2px 0px 2px rgba(92, 37, 19, 1)",
+        "--sjs-font-editorfont-color": "rgba(18, 18, 6, 1)",
+        "--sjs-font-editorfont-placeholdercolor": "rgba(18, 18, 6, 1)",
+        "--sjs-border-default": "rgba(0, 0, 0, 1)",
+        "--sjs-border-light": "rgba(0, 0, 0, 0.2)",
+        "--sjs-general-backcolor": "rgba(255, 255, 255, 1)",
+        "--sjs-general-backcolor-dark": "rgba(255, 216, 77, 1)",
+        "--sjs-general-backcolor-dim-light": "rgba(255, 216, 77, 1)",
+        "--sjs-general-backcolor-dim-dark": "rgba(255, 216, 77, 1)",
+        "--sjs-general-forecolor": "rgba(0, 0, 0, 1)",
+        "--sjs-general-forecolor-light": "rgba(0, 0, 0, 1)",
+        "--sjs-general-dim-forecolor": "rgba(0, 0, 0, 1)",
+        "--sjs-general-dim-forecolor-light": "rgba(0, 0, 0, 1)",
+        "--sjs-secondary-backcolor": "rgba(255, 152, 20, 1)",
+        "--sjs-secondary-backcolor-light": "rgba(255, 152, 20, 0.1)",
+        "--sjs-secondary-backcolor-semi-light": "rgba(255, 152, 20, 0.25)",
+        "--sjs-secondary-forecolor": "rgba(255, 255, 255, 1)",
+        "--sjs-secondary-forecolor-light": "rgba(255, 255, 255, 0.25)",
+        "--sjs-shadow-small-reset": "0px 0px 0px 0px rgba(92, 37, 19, 1),0px 0px 0px 0px rgba(191, 73, 4, 0.8)",
+        "--sjs-shadow-medium": "0px 0px 0px 2px rgba(0, 0, 0, 1)",
+        "--sjs-shadow-large": "0px 6px 0px 0px rgba(0, 0, 0, 1)",
+        "--sjs-shadow-inner-reset": "0px 0px 0px 0px rgba(92, 37, 19, 1),0px 0px 0px 0px rgba(92, 37, 19, 1)",
+        "--sjs-border-inside": "rgba(0, 0, 0, 0.16)",
+        "--sjs-special-red-forecolor": "rgba(255, 255, 255, 1)",
+        "--sjs-special-green": "rgba(25, 179, 148, 1)",
+        "--sjs-special-green-light": "rgba(25, 179, 148, 0.1)",
+        "--sjs-special-green-forecolor": "rgba(255, 255, 255, 1)",
+        "--sjs-special-blue": "rgba(67, 127, 217, 1)",
+        "--sjs-special-blue-light": "rgba(67, 127, 217, 0.1)",
+        "--sjs-special-blue-forecolor": "rgba(255, 255, 255, 1)",
+        "--sjs-special-yellow": "rgba(255, 152, 20, 1)",
+        "--sjs-special-yellow-light": "rgba(255, 152, 20, 0.1)",
+        "--sjs-special-yellow-forecolor": "rgba(255, 255, 255, 1)",
+        "--sjs-article-font-xx-large-textDecoration": "none",
+        "--sjs-article-font-xx-large-fontWeight": "700",
+        "--sjs-article-font-xx-large-fontStyle": "normal",
+        "--sjs-article-font-xx-large-fontStretch": "normal",
+        "--sjs-article-font-xx-large-letterSpacing": "0",
+        "--sjs-article-font-xx-large-lineHeight": "64px",
+        "--sjs-article-font-xx-large-paragraphIndent": "0px",
+        "--sjs-article-font-xx-large-textCase": "none",
+        "--sjs-article-font-x-large-textDecoration": "none",
+        "--sjs-article-font-x-large-fontWeight": "700",
+        "--sjs-article-font-x-large-fontStyle": "normal",
+        "--sjs-article-font-x-large-fontStretch": "normal",
+        "--sjs-article-font-x-large-letterSpacing": "0",
+        "--sjs-article-font-x-large-lineHeight": "56px",
+        "--sjs-article-font-x-large-paragraphIndent": "0px",
+        "--sjs-article-font-x-large-textCase": "none",
+        "--sjs-article-font-large-textDecoration": "none",
+        "--sjs-article-font-large-fontWeight": "700",
+        "--sjs-article-font-large-fontStyle": "normal",
+        "--sjs-article-font-large-fontStretch": "normal",
+        "--sjs-article-font-large-letterSpacing": "0",
+        "--sjs-article-font-large-lineHeight": "40px",
+        "--sjs-article-font-large-paragraphIndent": "0px",
+        "--sjs-article-font-large-textCase": "none",
+        "--sjs-article-font-medium-textDecoration": "none",
+        "--sjs-article-font-medium-fontWeight": "700",
+        "--sjs-article-font-medium-fontStyle": "normal",
+        "--sjs-article-font-medium-fontStretch": "normal",
+        "--sjs-article-font-medium-letterSpacing": "0",
+        "--sjs-article-font-medium-lineHeight": "32px",
+        "--sjs-article-font-medium-paragraphIndent": "0px",
+        "--sjs-article-font-medium-textCase": "none",
+        "--sjs-article-font-default-textDecoration": "none",
+        "--sjs-article-font-default-fontWeight": "400",
+        "--sjs-article-font-default-fontStyle": "normal",
+        "--sjs-article-font-default-fontStretch": "normal",
+        "--sjs-article-font-default-letterSpacing": "0",
+        "--sjs-article-font-default-lineHeight": "28px",
+        "--sjs-article-font-default-paragraphIndent": "0px",
+        "--sjs-article-font-default-textCase": "none",
+        "--sjs-general-backcolor-dim": "#F2AE72",
+        "--sjs-primary-backcolor": "#5C2513",
+        "--sjs-primary-backcolor-dark": "rgba(175, 70, 36, 1)",
+        "--sjs-primary-backcolor-light": "rgba(347, 140, 72, 1)",
+        "--sjs-primary-forecolor": "rgba(255, 255, 255, 1)",
+        "--sjs-primary-forecolor-light": "rgba(255, 255, 255, 0.25)",
+        "--sjs-special-red": "rgba(229, 221, 210, 1)",
+        "--sjs-special-red-light": "rgba(191, 73, 4, 0.8)",
+        "--sjs-font-headertitle-family": "Brush Script MT, cursive",
+        "--sjs-font-headertitle-weight": "600",
+        "--sjs-font-headertitle-color": "rgba(18, 18, 6, 1)",
+        "--sjs-font-headerdescription-color": "rgba(18, 18, 6, 1)",
+        "--sjs-header-backcolor": "transparent"
+    },
+    "header": {
+        "titlePositionX": "center",
+        "titlePositionY": "middle",
+        "descriptionPositionX": "center"
+    },
+    "headerView": "basic"
+}
+
+const tropicalTheme = {
+    "themeName": "coffee",
+    "colorPalette": "light",
+    "isPanelless": false,
+    "backgroundImage": "",
+    "backgroundOpacity": 1,
+    "backgroundImageAttachment": "scroll",
+    "backgroundImageFit": "cover",
+    "cssVariables": {
+        "--sjs-editorpanel-backcolor": "rgba(242, 174, 114, 1)",
+        "--sjs-editorpanel-hovercolor": "rgba(220, 83, 4, 1)",
+        "--sjs-questionpanel-backcolor": "rgba(229, 221, 210, 1)",
+        "--sjs-questionpanel-hovercolor": "rgba(242, 174, 114, 0.5)",
+        "--sjs-questionpanel-cornerRadius": "10px",
+        "--sjs-corner-radius": "10px",
+        "--sjs-base-unit": "8px",
+        "--sjs-font-pagetitle-family": "Brush Script MT, cursive",
+        "--sjs-font-pagetitle-color": "rgba(92, 37, 19, 1)",
+        "--sjs-font-pagedescription-color": "rgba(18, 18, 6, 1)",
+        "--sjs-shadow-small": "0px 0px 0px 3px rgba(92, 37, 19, 1),-5px 5px 0px 5px rgba(191, 73, 4, 0.8)",
+        "--sjs-font-questiontitle-color": "rgba(18, 18, 6, 1)",
+        "--sjs-font-questiondescription-size": "12px",
+        "--sjs-font-questiondescription-color": "rgba(18, 18, 6, 1)",
+        "--sjs-shadow-inner": "0px 0px 0px 2px rgba(92, 37, 19, 1),0px -2px 0px 2px rgba(92, 37, 19, 1)",
+        "--sjs-font-editorfont-color": "rgba(18, 18, 6, 1)",
+        "--sjs-font-editorfont-placeholdercolor": "rgba(18, 18, 6, 1)",
+        "--sjs-border-default": "rgba(0, 0, 0, 1)",
+        "--sjs-border-light": "rgba(0, 0, 0, 0.2)",
+        "--sjs-general-backcolor": "rgba(255, 255, 255, 1)",
+        "--sjs-general-backcolor-dark": "rgba(255, 216, 77, 1)",
+        "--sjs-general-backcolor-dim-light": "rgba(255, 216, 77, 1)",
+        "--sjs-general-backcolor-dim-dark": "rgba(255, 216, 77, 1)",
+        "--sjs-general-forecolor": "rgba(0, 0, 0, 1)",
+        "--sjs-general-forecolor-light": "rgba(0, 0, 0, 1)",
+        "--sjs-general-dim-forecolor": "rgba(0, 0, 0, 1)",
+        "--sjs-general-dim-forecolor-light": "rgba(0, 0, 0, 1)",
+        "--sjs-secondary-backcolor": "rgba(255, 152, 20, 1)",
+        "--sjs-secondary-backcolor-light": "rgba(255, 152, 20, 0.1)",
+        "--sjs-secondary-backcolor-semi-light": "rgba(255, 152, 20, 0.25)",
+        "--sjs-secondary-forecolor": "rgba(255, 255, 255, 1)",
+        "--sjs-secondary-forecolor-light": "rgba(255, 255, 255, 0.25)",
+        "--sjs-shadow-small-reset": "0px 0px 0px 0px rgba(92, 37, 19, 1),0px 0px 0px 0px rgba(191, 73, 4, 0.8)",
+        "--sjs-shadow-medium": "0px 0px 0px 2px rgba(0, 0, 0, 1)",
+        "--sjs-shadow-large": "0px 6px 0px 0px rgba(0, 0, 0, 1)",
+        "--sjs-shadow-inner-reset": "0px 0px 0px 0px rgba(92, 37, 19, 1),0px 0px 0px 0px rgba(92, 37, 19, 1)",
+        "--sjs-border-inside": "rgba(0, 0, 0, 0.16)",
+        "--sjs-special-red-forecolor": "rgba(255, 255, 255, 1)",
+        "--sjs-special-green": "rgba(25, 179, 148, 1)",
+        "--sjs-special-green-light": "rgba(25, 179, 148, 0.1)",
+        "--sjs-special-green-forecolor": "rgba(255, 255, 255, 1)",
+        "--sjs-special-blue": "rgba(67, 127, 217, 1)",
+        "--sjs-special-blue-light": "rgba(67, 127, 217, 0.1)",
+        "--sjs-special-blue-forecolor": "rgba(255, 255, 255, 1)",
+        "--sjs-special-yellow": "rgba(255, 152, 20, 1)",
+        "--sjs-special-yellow-light": "rgba(255, 152, 20, 0.1)",
+        "--sjs-special-yellow-forecolor": "rgba(255, 255, 255, 1)",
+        "--sjs-article-font-xx-large-textDecoration": "none",
+        "--sjs-article-font-xx-large-fontWeight": "700",
+        "--sjs-article-font-xx-large-fontStyle": "normal",
+        "--sjs-article-font-xx-large-fontStretch": "normal",
+        "--sjs-article-font-xx-large-letterSpacing": "0",
+        "--sjs-article-font-xx-large-lineHeight": "64px",
+        "--sjs-article-font-xx-large-paragraphIndent": "0px",
+        "--sjs-article-font-xx-large-textCase": "none",
+        "--sjs-article-font-x-large-textDecoration": "none",
+        "--sjs-article-font-x-large-fontWeight": "700",
+        "--sjs-article-font-x-large-fontStyle": "normal",
+        "--sjs-article-font-x-large-fontStretch": "normal",
+        "--sjs-article-font-x-large-letterSpacing": "0",
+        "--sjs-article-font-x-large-lineHeight": "56px",
+        "--sjs-article-font-x-large-paragraphIndent": "0px",
+        "--sjs-article-font-x-large-textCase": "none",
+        "--sjs-article-font-large-textDecoration": "none",
+        "--sjs-article-font-large-fontWeight": "700",
+        "--sjs-article-font-large-fontStyle": "normal",
+        "--sjs-article-font-large-fontStretch": "normal",
+        "--sjs-article-font-large-letterSpacing": "0",
+        "--sjs-article-font-large-lineHeight": "40px",
+        "--sjs-article-font-large-paragraphIndent": "0px",
+        "--sjs-article-font-large-textCase": "none",
+        "--sjs-article-font-medium-textDecoration": "none",
+        "--sjs-article-font-medium-fontWeight": "700",
+        "--sjs-article-font-medium-fontStyle": "normal",
+        "--sjs-article-font-medium-fontStretch": "normal",
+        "--sjs-article-font-medium-letterSpacing": "0",
+        "--sjs-article-font-medium-lineHeight": "32px",
+        "--sjs-article-font-medium-paragraphIndent": "0px",
+        "--sjs-article-font-medium-textCase": "none",
+        "--sjs-article-font-default-textDecoration": "none",
+        "--sjs-article-font-default-fontWeight": "400",
+        "--sjs-article-font-default-fontStyle": "normal",
+        "--sjs-article-font-default-fontStretch": "normal",
+        "--sjs-article-font-default-letterSpacing": "0",
+        "--sjs-article-font-default-lineHeight": "28px",
+        "--sjs-article-font-default-paragraphIndent": "0px",
+        "--sjs-article-font-default-textCase": "none",
+        "--sjs-general-backcolor-dim": "#F2AE72",
+        "--sjs-primary-backcolor": "#5C2513",
+        "--sjs-primary-backcolor-dark": "rgba(175, 70, 36, 1)",
+        "--sjs-primary-backcolor-light": "rgba(347, 140, 72, 1)",
+        "--sjs-primary-forecolor": "rgba(255, 255, 255, 1)",
+        "--sjs-primary-forecolor-light": "rgba(255, 255, 255, 0.25)",
+        "--sjs-special-red": "rgba(229, 221, 210, 1)",
+        "--sjs-special-red-light": "rgba(191, 73, 4, 0.8)",
+        "--sjs-font-headertitle-family": "Brush Script MT, cursive",
+        "--sjs-font-headertitle-weight": "600",
+        "--sjs-font-headertitle-color": "rgba(18, 18, 6, 1)",
+        "--sjs-font-headerdescription-color": "rgba(18, 18, 6, 1)",
+        "--sjs-header-backcolor": "transparent"
+    },
+    "header": {
+        "titlePositionX": "center",
+        "titlePositionY": "middle",
+        "descriptionPositionX": "center"
+    },
+    "headerView": "basic"
+}
+
+export { coffeeTheme, tropicalTheme};
